+++ packages/hightide-native/src/global-contexts/hightide-config/HightideConfigUtils.ts
@@ -1,14 +1,14 @@
 import type { SupportedThemesConfig } from '@helpwave/hightide-utils/context'
 
-import type { LocaleInformation } from '@/src/global-contexts/localization/forward-exports'
+import type { SupportedLocalesConfig } from '@/src/global-contexts/localization/forward-exports'
 import { themes } from '@/src/theme/themes/nativeThemes'
 import type { Theme } from '@/src/theme/types/theme'
 
 export const HightideConfigUtils = {
-  defaultSupportedLocales: [
-    { locale: 'de-DE', localName: 'Deutsch' },
-    { locale: 'en-US', localName: 'English (US)' },
-  ] as const satisfies readonly LocaleInformation[],
+  defaultSupportedLocales: {
+    'de-DE': { localName: 'Deutsch' },
+    'en-US': { localName: 'English (US)' },
+  } as const satisfies SupportedLocalesConfig,
   defaultSupportedThemes: {
     light: {
       nameTranslations: {